movies_admin/requirements/dev.txt
-r base.txt

# Конфигурируемая панель с информацией для отладки проекта
django-debug-toolbar==2.2

# генерация тестовых данных
factory-boy==3.2.0